ms in long

+++ android/src/com/google/iamnotok/EmergencyNotificationService.java
@@ -60,7 +60,7 @@ public class EmergencyNotificationService extends Service {
 
 	/** Default time allowed for user to cancel the emergency response. */
 	private static long DEFAULT_WAIT_TO_CANCEL_MS = 10000;
-	private static final int DEFAULT_WAIT_BETWEEN_MESSAGES_MS = 5 * 60 * 1000;
+	private static final long DEFAULT_WAIT_BETWEEN_MESSAGES_MS = 5 * 60 * 1000;
 
 
 	private int mNotificationID = 0;
